ソースを見やすいように改良
