+++ fruit-get/fruit.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void fruit_move(void);
+int fruit_get(int score, int x);
+void fruit_gene(void);
+void show(int cx, int cy, int score);
+void fruit_reset(void);
