+++ fruit-get/oprate.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void opening(void);
+int ending(int op);
+void end_sen(void);
