+++ fruit-get/player.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int player_move(int cx, int key);
